✨ Password Reset Request Form, some refactoring of user routes
