+++ app/client-app/src/app/pages/user/Register/index.js
@@ -1,129 +1,23 @@
-import React, { useState } from "react";
+import React from "react";
 import { Page } from "components/core";
 import { postObjectAsFormData } from "js-forms";
 import { Formik, Form, Field } from "formik";
-import {
-  Stack,
-  Button,
-  Flex,
-  Popover,
-  PopoverArrow,
-  PopoverBody,
-  PopoverContent,
-  PopoverTrigger,
-  Link,
-  Alert,
-  AlertIcon,
-} from "@chakra-ui/core";
+import { Stack, Button, Flex, Alert, AlertIcon } from "@chakra-ui/core";
 import LightHeading from "components/core/LightHeading";
 import FormikInput from "components/core/FormikInput";
 import validationSchema from "./validation";
 import { navigate } from "@reach/router";
 import { useQueryStringViewModel } from "hooks/useQueryString";
 import { useServerConfig } from "api/config";
-import Error from "../Error";
+import Error from "app/pages/Error";
 import ErrorsAlert from "components/core/ErrorsAlert";
 import { ApprovalRequired, EmailConfirmationRequired } from "./alerts";
 import {
   REQUIRES_APPROVAL,
   REQUIRES_EMAIL_CONFIRMATION,
 } from "constants/account-states";
-
-// TODO: this will later want to be reusable for email change
-const EmailFieldGroup = ({ initialHidden }) => {
-  const [hidden, setHidden] = useState(initialHidden);
-  const handleFocus = () => setHidden(false);
-
-  return (
-    <>
-      <Field name="Email">
-        {(rp) => (
-          <FormikInput
-            {...rp}
-            label="Email Address"
-            placeholder="john.smith@example.com"
-            isRequired
-            onFocus={handleFocus}
-          />
-        )}
-      </Field>
-
-      <Flex hidden={hidden}>
-        <Field name="EmailConfirm">
-          {(rp) => (
-            <FormikInput
-              {...rp}
-              label="Confirm Email Address"
-              placeholder="john.smith@example.com"
-              isRequired
-            />
-          )}
-        </Field>
-      </Flex>
-    </>
-  );
-};
-
-// TODO: this will later want to be reusable for password reset
-const PasswordFieldGroup = ({ initialHidden }) => {
-  const [hidden, setHidden] = useState(initialHidden);
-  const handleFocus = () => setHidden(false);
-
-  const tip = (
-    <Popover returnFocusOnClose={false} usePortal>
-      <PopoverTrigger>
-        <Link color="blue.500" href="#">
-          Password Requirements
-        </Link>
-      </PopoverTrigger>
-      <PopoverContent bg="gray.300" borderColor="gray.400">
-        <PopoverArrow />
-        <PopoverBody pl={8}>
-          <ul>
-            <li>Passwords must be at least 6 characters.</li>
-            <li>Passwords must have at least one digit ('0' - '9').</li>
-            <li>Passwords must have at least one uppercase ('A' - 'Z').</li>
-            <li>
-              Passwords must have at least one non alphanumeric character.
-            </li>
-          </ul>
-        </PopoverBody>
-      </PopoverContent>
-    </Popover>
-  );
-
-  return (
-    <>
-      <Field name="Password">
-        {(rp) => (
-          <FormikInput
-            {...rp}
-            label={rp.field.name}
-            placeholder={rp.field.name}
-            isRequired
-            isPassword
-            fieldTip={tip}
-            onFocus={handleFocus}
-          />
-        )}
-      </Field>
-
-      <Flex hidden={hidden}>
-        <Field name="PasswordConfirm">
-          {(rp) => (
-            <FormikInput
-              {...rp}
-              label="Confirm Password"
-              placeholder="Password"
-              isRequired
-              isPassword
-            />
-          )}
-        </Field>
-      </Flex>
-    </>
-  );
-};
+import EmailFieldGroup from "../components/EmailFieldGroup";
+import PasswordFieldGroup from "../components/PasswordFieldGroup";
 
 const Feedback = ({ errors, accountState, Email }) => {
   switch (accountState) {
@@ -185,6 +79,8 @@ const Register = () => {
               Fullname,
               Email,
               EmailConfirm,
+              Password: "",
+              PasswordConfirm: "",
             }}
             onSubmit={handleSubmit}
             validationSchema={validationSchema}
